Add lifetime scope support for SignalR Hubs
Added a RegisterHubWithLifetimeScope<T> method that creates a
dynamic proxy wrapper around T. The proxy class is then registered as
the T component. When a T component is resolved, it creates a new
implicit lifetime scope that is destroyed via interceptor when SignalR
disposes the Hub. This causes all the lifetime-scoped IDisposable
dependencies of the Hub to be cleaned up at the same time.